hay 120 bytes q no se xq no se liberan en debug version c

+++ trunk/debug-version/debug-C/dummy_audio.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <math.h>
 
-typedef float  dummy_audio_size;
+typedef int    dummy_audio_size;
 typedef float* dummy_audio_buf;
 
 typedef struct _dummy_audio{
